add redirect back to houses route after user remove themselves from a house that only contain themselves
